app/backend/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-cors==4.0.0
pymysql==1.1.0
bcrypt==4.1.2
python-dotenv==1.0.0
gunicorn==21.2.0
marshmallow==3.20.1
